pyproject.toml
[project]
name = "telegram-scraper"
version = "0.1.0"
description = "Telegram channel scraper with optional translation, title generation, Mongo storage, and webhook forwarding."
readme = "README.md"
requires-python = ">=3.10,<3.15"
license = "MIT"
authors = [
  { name = "telegram-scraper contributors" }
]
keywords = ["telegram", "scraper", "ai", "mongodb", "telethon"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Internet",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "telethon>=1.42.0,<2.0.0",
  "python-dotenv>=1.2.1,<2.0.0",
  "requests>=2.32.5,<3.0.0",
  "urllib3>=2.6.2,<3.0.0",
]

[project.optional-dependencies]
genai = ["google-genai>=1.56.0,<2.0.0"]
local-ai = ["torch>=2.9.1,<3.0.0", "transformers>=4.57.3,<5.0.0"]
mongo = ["pymongo>=4.15.5,<5.0.0"]
all = [
  "google-genai>=1.56.0,<2.0.0",
  "pymongo>=4.15.5,<5.0.0",
  "torch>=2.9.1,<3.0.0",
  "transformers>=4.57.3,<5.0.0",
]

[project.scripts]
telegram-scrape = "telegram_intel_scraper.main:main"

[tool.poetry]
packages = [
  { include = "telegram_intel_scraper", from = "src" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
ruff = "^0.14.0"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120

[tool.coverage.run]
source = ["src/telegram_intel_scraper"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
